LayoutTests/fast/css/cssText-shorthand-expected.txt

This test ensures WebKit uses shorthand notations for cssText

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS normalizeCssText(element.style.cssText) is "border: 1px"
PASS normalizeCssText(element.style.cssText) is "border: 1px solid red"
PASS normalizeCssText(element.style.cssText) is "border: 1px red"
PASS normalizeCssText(element.style.cssText) is "border: red"
PASS normalizeCssText(element.style.cssText) is "border: 1px"
PASS normalizeCssText(element.style.cssText) is "border-bottom-width: 3px; border-left-width: 4px; border-right-width: 2px; border-top-width: 1px"
PASS normalizeCssText(element.style.cssText) is "border-bottom-width: 1px; border-left-width: 1px; border-right-width: 1px; border-top-width: 2px"
PASS normalizeCssText(element.style.cssText) is "border-bottom-width: 1px; border-left-width: 1px; border-right-width: 1px; border-top-width: 1px !important"
PASS normalizeCssText(element.style.cssText) is "border-bottom-width: 1px; border-left-width: 1px; border-right-width: 1px; border-top-color: red; border-top-width: 1px"
PASS normalizeCssText(element.style.cssText) is "border: dotted"
PASS normalizeCssText(element.style.cssText) is "border-spacing: 1px 2px"
PASS normalizeCssText(element.style.cssText) is "font-family: sans-serif; font-size: 3em; font-style: italic; font-weight: bold; line-height: 2em"
PASS normalizeCssText(element.style.cssText) is "list-style: circle inside"
PASS normalizeCssText(element.style.cssText) is "margin: 1px 2px 3px 4px"
PASS normalizeCssText(element.style.cssText) is "outline: blue dotted 2px"
FAIL normalizeCssText(element.style.cssText) should be overflow: scroll hidden. Was overflow-x: scroll; overflow-y: hidden.
PASS normalizeCssText(element.style.cssText) is "padding: 1px 2px 3px 4px"
PASS successfullyParsed is true

TEST COMPLETE
